fix warnings

+++ src/client/SetupManager.cpp
@@ -11,23 +11,22 @@ int *SetupManager::_grid;
  */
 SetupManager::SetupManager() {
   // generate ships
-  _ships_placed.emplace_back(Ship(5, Coordinate{-1,-1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
-  _ships_placed.emplace_back(Ship(4, Coordinate{-1,-1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
-  _ships_placed.emplace_back(Ship(3, Coordinate{-1,-1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
-  _ships_placed.emplace_back(Ship(3, Coordinate{-1,-1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
-  _ships_placed.emplace_back(Ship(2, Coordinate{-1,-1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
+  _ships_placed.emplace_back(Ship(5, Coordinate{-1, -1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
+  _ships_placed.emplace_back(Ship(4, Coordinate{-1, -1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
+  _ships_placed.emplace_back(Ship(3, Coordinate{-1, -1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
+  _ships_placed.emplace_back(Ship(3, Coordinate{-1, -1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
+  _ships_placed.emplace_back(Ship(2, Coordinate{-1, -1}, Ship::Orientation::Horizontal, uuid::generateRandomUuid()));
   _selectedShip = nullptr;
 
   _grid = new int[100];
 
   // init grid to 0
-  for(int i = 0; i < 10; ++i) {
-        for(int j = 0; j < 10; ++j) {
-          _grid[i*10+j] = 0;
-        }
+  for (int i = 0; i < 10; ++i) {
+    for (int j = 0; j < 10; ++j) {
+      _grid[i * 10 + j] = 0;
+    }
   }
 }
-
 
 /**
  * @brief function places ship on grid and updates ship position
@@ -36,8 +35,8 @@ SetupManager::SetupManager() {
  * @return true if ship was placed successfully and false if ship couldn't be placed
  */
 bool SetupManager::placeShip(wxPoint &position, Ship *ship) {
-  if(ship == nullptr)
-        return false;
+  if (ship == nullptr)
+    return false;
 
   // check in grid if position is valid
   const int length      = ship->getLength();
@@ -48,19 +47,19 @@ bool SetupManager::placeShip(wxPoint &position, Ship *ship) {
 
   // if cursor is too far right or bottom, place ship at right or bottom edge
   if ((orientation == Ship::Orientation::Horizontal && cellX + length > 10)) {
-          cellX = std::max(10 - length, 0);
+    cellX = std::max(10 - length, 0);
   }
   if ((orientation == Ship::Orientation::Vertical && cellY + length > 10)) {
-          cellY = std::max(10 - length, 0);
+    cellY = std::max(10 - length, 0);
   }
 
   // check if ship would overlap with another ship
   bool overlap = false;
   for (int i = 0; i < length; ++i) {
-        if (orientation == Ship::Orientation::Horizontal)
-          overlap = overlap || _grid[(cellX + i) * 10 + cellY] != 0;
-        else
-          overlap = overlap || _grid[(cellX) * 10 + cellY + i] != 0;
+    if (orientation == Ship::Orientation::Horizontal)
+      overlap = overlap || _grid[(cellX + i) * 10 + cellY] != 0;
+    else
+      overlap = overlap || _grid[(cellX)*10 + cellY + i] != 0;
   }
   if (overlap) {
     LOG("invalid placement. try again.");
@@ -68,38 +67,36 @@ bool SetupManager::placeShip(wxPoint &position, Ship *ship) {
   }
 
   // get idx of ship in _ships_placed
-  for(int i = 0; i < _ships_placed.size(); ++i) {
-        if(_ships_placed[i].getId() == ship->getId()) {
-          // update ship position
-          _ships_placed[i].setPosition(Coordinate{cellX, cellY});
-          for(int j = 0; j < length; ++j) { // set ship id in grid array
-            if(orientation == Ship::Orientation::Horizontal)
-              _grid[(cellX+j)*10 + cellY] = i + 1;
-            else
-              _grid[(cellX)*10 + cellY+ j] = i + 1;
-          }
-          break;
-        }
+  for (unsigned int i = 0; i < _ships_placed.size(); ++i) {
+    if (_ships_placed[i].getId() == ship->getId()) {
+      // update ship position
+      _ships_placed[i].setPosition(Coordinate{cellX, cellY});
+      for (int j = 0; j < length; ++j) { // set ship id in grid array
+        if (orientation == Ship::Orientation::Horizontal)
+          _grid[(cellX + j) * 10 + cellY] = static_cast<int>(i + 1);
+        else
+          _grid[(cellX)*10 + cellY + j] = static_cast<int>(i + 1);
+      }
+      break;
+    }
   }
 
   // remove Ship button from UI
-  auto ship_idx = _grid[cellX*10 + cellY] - 1; // idx in [0,4]
-  auto *shipBtn = GameController::getSetupPanel()->getShipButton(ship_idx);
-  if(shipBtn != nullptr) // should never be nullptr but just in case
-        shipBtn->Disable();
-
+  auto  ship_idx = _grid[cellX * 10 + cellY] - 1; // idx in [0,4]
+  auto *shipBtn  = GameController::getSetupPanel()->getShipButton(ship_idx);
+  if (shipBtn != nullptr) // should never be nullptr but just in case
+    shipBtn->Disable();
 
   // print grid (debug)
   for (int i = 0; i < 10; ++i) {
-        for (int j = 0; j < 10; ++j) {
-          std::cout << _grid[i + j * 10] << " ";
-        }
-        std::cout << std::endl;
+    for (int j = 0; j < 10; ++j) {
+      std::cout << _grid[i + j * 10] << " ";
+    }
+    std::cout << std::endl;
   }
 
   return true;
 }
-
 
 /**
  * @brief getter for _grid
@@ -109,15 +106,14 @@ int *SetupManager::getGrid() {
   return _grid;
 }
 
-
 /**
  * @brief checks if all ships have been placed (= no longer at initial position)
  * @return true if all ships have been placed, false otherwise
  */
 bool SetupManager::placedAllShips() {
   bool placed = true;
-  for(int i = 0; i < _ships_placed.size(); ++i) {
-        placed = placed && _ships_placed[i].getPosition() != Coordinate{-1,-1};
+  for (unsigned int i = 0; i < _ships_placed.size(); ++i) {
+    placed = placed && _ships_placed[i].getPosition() != Coordinate{-1, -1};
   }
   return placed;
 }
